requirements.txt
fastmcp==2.13.1
fastapi==0.121.3
httpx==0.28.1
python-json-logger==4.0.0

# Logging / telemetry
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-instrumentation-httpx==0.48b0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
